Complete part 6
Part 6.1 was really easy

Part 6.2 was difficult for a little bit as you can't just parse with a regex,
you have to realize that there is insufficient information to do simple line by
line parsing (you don't know if excess spaces apply to the first grouping or the
second grouping for positioning with columns.  So I spent a lot of time
trying to avoid parsing everything at once.